requirements.txt
dbus-python==1.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0